+++ src/provider/client/index.ts
@@ -14,6 +14,11 @@ import { QualysIntegrationConfig } from '../../types';
 import { executeAPIRequest } from './request';
 import {
   assets,
+  ClientDelayedRequestEvent,
+  ClientEvents,
+  ClientRequestEvent,
+  ClientResponseEvent,
+  QWebHostId,
   RateLimitConfig,
   RateLimitState,
   RetryConfig,
@@ -32,9 +37,9 @@ export * from './types';
 
 export type ResourceIteratee<T> = (each: T) => Promise<void> | void;
 
-const DEFAULT_RETRY_CONFIG: RetryConfig = {
+export const DEFAULT_RETRY_CONFIG: RetryConfig = {
   maxAttempts: 5,
-  noRetry: [400, 401, 403],
+  noRetry: [400, 401, 403, 413],
 };
 
 /**
@@ -43,7 +48,7 @@ const DEFAULT_RETRY_CONFIG: RetryConfig = {
  * @see https://www.qualys.com/docs/qualys-api-limits.pdf for details on
  * subscription level rate limits.
  */
-const STANDARD_RATE_LIMIT_STATE: RateLimitState = {
+export const STANDARD_RATE_LIMIT_STATE: RateLimitState = {
   limit: 300,
   limitRemaining: 300,
   limitWindowSeconds: 60 * 60,
@@ -52,7 +57,7 @@ const STANDARD_RATE_LIMIT_STATE: RateLimitState = {
   concurrencyRunning: 0,
 };
 
-const DEFAULT_RATE_LIMIT_CONFIG: RateLimitConfig = {
+export const DEFAULT_RATE_LIMIT_CONFIG: RateLimitConfig = {
   responseCode: 409,
   maxAttempts: 5,
   reserveLimit: 30,
@@ -89,7 +94,7 @@ export type QualysAPIClientConfig = {
 };
 
 export class QualysAPIClient {
-  public events: EventEmitter;
+  private events: EventEmitter;
 
   private config: QualysIntegrationConfig;
   private retryConfig: RetryConfig;
@@ -107,8 +112,22 @@ export class QualysAPIClient {
     this.config = config;
     this.retryConfig = { ...DEFAULT_RETRY_CONFIG, ...retryConfig };
     this.rateLimitConfig = { ...DEFAULT_RATE_LIMIT_CONFIG, ...rateLimitConfig };
-    this.rateLimitState = rateLimitState || { ...STANDARD_RATE_LIMIT_STATE };
+    this.rateLimitState = rateLimitState || STANDARD_RATE_LIMIT_STATE;
     this.events = new EventEmitter();
+  }
+
+  public onRequest(eventHandler: (event: ClientRequestEvent) => void): void {
+    this.events.on(ClientEvents.REQUEST, eventHandler);
+  }
+
+  public onDelayedRequest(
+    eventHandler: (event: ClientDelayedRequestEvent) => void,
+  ): void {
+    this.events.on(ClientEvents.DELAYED_REQUEST, eventHandler);
+  }
+
+  public onResponse(eventHandler: (event: ClientResponseEvent) => void): void {
+    this.events.on(ClientEvents.RESPONSE, eventHandler);
   }
 
   public async verifyAuthentication(): Promise<void> {
@@ -299,14 +318,8 @@ export class QualysAPIClient {
   }
 
   /**
-   * Answers the complete set of scanned host IDs provided by the Qualys VMDR
+   * Answers the complete set of scanned `QWebHostId`s provided by the Qualys VMDR
    * module. This does not include hosts that have never been scanned.
-   *
-   * There are three IDs in Qualys. The IDs returned by this API are the VM
-   * module "QWEB" host IDs.
-   *
-   * @see https://qualys-secure.force.com/discussions/s/article/000006216 to
-   * understand the difference.
    *
    * Fetches all IDs in a single request. This is documented by Qualys as a best
    * practice for an implementation that will parallelize ingestion of other
@@ -332,9 +345,66 @@ export class QualysAPIClient {
 
     const responseText = await response.text();
     const jsonFromXml = xmlParser.parse(responseText);
-    return toArray(jsonFromXml.HOST_LIST_OUTPUT?.RESPONSE?.ID_SET).map(
-      (e) => e.ID,
+    return toArray(jsonFromXml.HOST_LIST_OUTPUT?.RESPONSE?.ID_SET?.ID);
+  }
+
+  /**
+   * Answers the complete set of scanned host IDs provided by the Qualys VMDR
+   * module. This does not include hosts that have never been scanned.
+   *
+   * @param iteratee receives each page of host ID values
+   * @param options optional values for pagination
+   */
+  public async iterateScannedHostIds(
+    iteratee: ResourceIteratee<QWebHostId[]>,
+    options?: {
+      pageSize: number;
+    },
+  ): Promise<void> {
+    type ListHostIdsResponse = {
+      nextUrl?: string;
+      hostIds: QWebHostId[];
+    };
+
+    const buildHostIdsResponse = async (
+      response: Response,
+    ): Promise<ListHostIdsResponse> => {
+      const responseText = await response.text();
+      const jsonFromXml = xmlParser.parse(responseText);
+
+      const hostList = jsonFromXml.HOST_LIST_OUTPUT?.RESPONSE?.HOST_LIST;
+      const idSet = jsonFromXml.HOST_LIST_OUTPUT?.RESPONSE?.ID_SET;
+
+      return {
+        hostIds: hostList
+          ? toArray(hostList.HOST).map((host) => host.ID)
+          : toArray(idSet.ID),
+        nextUrl: jsonFromXml.HOST_LIST_OUTPUT?.RESPONSE?.WARNING?.URL,
+      };
+    };
+
+    const endpoint = '/api/2.0/fo/asset/host/';
+    const response = await this.executeAuthenticatedAPIRequest(
+      this.qualysUrl(endpoint, {
+        action: 'list',
+        details: 'None',
+        truncation_limit: options?.pageSize || 500,
+      }),
+      { method: 'GET' },
     );
+
+    let hostIdsResponse = await buildHostIdsResponse(response);
+    await iteratee(hostIdsResponse.hostIds);
+
+    while (hostIdsResponse.nextUrl) {
+      const response = await this.executeAuthenticatedAPIRequest(
+        this.qualysUrl(hostIdsResponse.nextUrl),
+        { method: 'GET' },
+      );
+
+      hostIdsResponse = await buildHostIdsResponse(response);
+      await iteratee(hostIdsResponse.hostIds);
+    }
   }
 
   /**
@@ -347,10 +417,10 @@ export class QualysAPIClient {
    * @param hostIds a set of identified QWEB host IDs
    */
   public async iterateHostDetails(
-    hostIds: number[],
+    hostIds: QWebHostId[],
     iteratee: ResourceIteratee<assets.HostAsset>,
   ): Promise<void> {
-    const fetchHostDetails = async (ids: number[]) => {
+    const fetchHostDetails = async (ids: QWebHostId[]) => {
       const endpoint = '/qps/rest/2.0/search/am/hostasset';
 
       const body = `
@@ -407,7 +477,7 @@ export class QualysAPIClient {
    *
    * @param hostId a QWEB host ID
    */
-  public async fetchHostDetails(hostId: number): Promise<assets.HostAsset> {
+  public async fetchHostDetails(hostId: QWebHostId): Promise<assets.HostAsset> {
     const endpoint = '/qps/rest/2.0/search/am/hostasset';
 
     const body = `
@@ -462,13 +532,13 @@ export class QualysAPIClient {
    * @param iteratee receives each host and its detections
    */
   public async iterateHostDetections(
-    hostIds: number[],
+    hostIds: QWebHostId[],
     iteratee: ResourceIteratee<{
       host: vmpc.DetectionHost;
       detections: vmpc.HostDetection[];
     }>,
   ): Promise<void> {
-    const fetchHostDetections = async (ids: number[]) => {
+    const fetchHostDetections = async (ids: QWebHostId[]) => {
       const endpoint = '/api/2.0/fo/asset/host/vm/detection/';
 
       const params = new URLSearchParams({
@@ -550,7 +620,7 @@ export class QualysAPIClient {
     }
   }
 
-  private async executeAuthenticatedAPIRequest(
+  public async executeAuthenticatedAPIRequest(
     info: RequestInfo,
     init: RequestInit,
   ): Promise<Response> {
@@ -575,11 +645,23 @@ export class QualysAPIClient {
       exec: () => fetch(info, init),
       retryConfig: this.retryConfig,
       rateLimitConfig: this.rateLimitConfig,
-      rateLimitState: this.rateLimitState,
+      rateLimitState: { ...this.rateLimitState },
     });
 
     // NOTE: This is NOT thread safe at this time.
     this.rateLimitState = apiResponse.rateLimitState;
+
+    if (!apiResponse.completed && apiResponse.request.retryable) {
+      const err = new Error(
+        `Could not complete request within ${apiResponse.request.totalAttempts} attempts!`,
+      );
+      Object.assign(err, {
+        statusText: apiResponse.statusText,
+        status: apiResponse.status,
+        code: apiResponse.status,
+      });
+      throw err;
+    }
 
     if (apiResponse.status >= 400) {
       const err = new Error(