+++ src/provider/client/request.ts
@@ -4,6 +4,7 @@ import { Response } from 'node-fetch';
 
 import {
   ClientDelayedRequestEvent,
+  ClientEvent,
   ClientEvents,
   ClientRequestEvent,
   ClientResponseEvent,
@@ -103,25 +104,47 @@ async function attemptAPIRequest(
   const toWaitSec = rateLimitState.toWaitSeconds;
   const tryAfter = Math.max(Date.now() + toWaitSec * 1000, tryAfterCooldown);
 
+  const requestEvent: ClientEvent = {
+    url: request.url,
+    retryConfig: request.retryConfig,
+    retryable: request.retryable,
+    retryAttempts: request.retryAttempts,
+    rateLimitConfig: request.rateLimitConfig,
+    rateLimitState: request.rateLimitState,
+    rateLimitedAttempts: request.rateLimitedAttempts,
+    totalAttempts: request.totalAttempts,
+  };
+
   const now = Date.now();
   if (tryAfter > now) {
     const delay = tryAfter - now;
-    emitDelayedRequestEvent(events, { ...request, delay });
+    emitDelayedRequestEvent(events, { ...requestEvent, delay });
     await Timeout.set(delay);
   }
 
-  emitRequestEvent(events, request);
+  emitRequestEvent(events, requestEvent);
 
   const response = await request.exec();
 
-  const completed = response.status >= 200 && response.status < 400;
-  const rateLimited = response.status === rateLimitConfig.responseCode;
   const responseRateLimitState = extractRateLimitHeaders(
     response,
     rateLimitConfig,
     rateLimitState,
   );
+
+  const completed = response.status >= 200 && response.status < 400;
+  const rateLimited = response.status === rateLimitConfig.responseCode;
+  const rateLimitedAttempts = rateLimited
+    ? request.rateLimitedAttempts + 1
+    : request.rateLimitedAttempts;
+
   const retryable = !retryConfig.noRetry.includes(response.status);
+  const retryAttempts =
+    !completed && !rateLimited
+      ? request.retryAttempts + 1
+      : request.retryAttempts;
+
+  const totalAttempts = request.totalAttempts + 1;
 
   const apiResponse: APIResponse = {
     request: {
@@ -129,14 +152,9 @@ async function attemptAPIRequest(
       completed,
       retryable,
       rateLimitState: responseRateLimitState,
-      totalAttempts: request.totalAttempts + 1,
-      retryAttempts:
-        !completed && !rateLimited
-          ? request.retryAttempts + 1
-          : request.retryAttempts,
-      rateLimitedAttempts: rateLimitState
-        ? request.rateLimitedAttempts + 1
-        : request.rateLimitedAttempts,
+      totalAttempts,
+      retryAttempts,
+      rateLimitedAttempts,
     },
     response,
     rateLimitState: responseRateLimitState,
@@ -149,11 +167,14 @@ async function attemptAPIRequest(
     url: request.url,
     status: response.status,
     statusText: response.statusText,
+    completed,
+    retryable,
     retryConfig: request.retryConfig,
-    retryAttempts: request.retryAttempts,
+    retryAttempts,
     rateLimitConfig: request.rateLimitConfig,
-    rateLimitState: request.rateLimitState,
-    rateLimitedAttempts: request.rateLimitedAttempts,
+    rateLimitState: responseRateLimitState,
+    rateLimitedAttempts,
+    totalAttempts,
   });
 
   return apiResponse;
@@ -176,9 +197,9 @@ function extractRateLimitHeaders(
 
   return {
     limit: limit ? Number(limit) : defaultState.limit,
-    limitWindowSeconds: limit
-      ? Number(response.headers.get('x-ratelimit-window-sec'))
-      : defaultState.limitWindowSeconds,
+    limitWindowSeconds:
+      Number(response.headers.get('x-ratelimit-window-sec')) ||
+      defaultState.limitWindowSeconds,
     limitRemaining: limit
       ? Number(response.headers.get('x-ratelimit-remaining'))
       : defaultState.limitRemaining,